almost working grid with clear
