+++ frontend/pages/courses/[course]/index.tsx
@@ -49,7 +49,7 @@ const QueuePage = (props: QueuePageProps) => {
             <Head>
                 <title>{`OHQ | ${course.department} ${course.courseCode}`}</title>
             </Head>
-            <Grid columns="equal" divided style={{ width: "100%" }} stackable>
+            <Grid divided style={{ width: "100%" }} stackable>
                 <CourseWrapper
                     course={course}
                     leadership={leadership}
